+++ src/config/config.types.d.ts
@@ -25,6 +25,7 @@ interface UserTwitchConfig {
     message: string;
     ignoreUsers: string[];
   };
+  maxMessageLength?: number;
   eventWebsocketUrl: string;
   eventSubscriptionUrl: string;
 }
@@ -71,11 +72,28 @@ interface UserTwitterConfig {
   announceLive: boolean;
 }
 
+interface UserDbConfig {
+  enabled: boolean;
+  type: 'postgres' | 'mysql' | 'sqlite';
+  host: string;
+  port?: number;
+  username: string;
+  password: string;
+  database: string;
+  migrationsRun: boolean;
+  extra?: {
+    ssl?: {
+      rejectUnauthorized: boolean;
+    };
+  };
+}
+
 interface UserDiscordConfig {
   enabled: boolean;
   botToken: string;
   announcementChannelId: string;
   galleryChannelId: string;
+  musicChannelId?: string;
   generalChannelId: string;
   modChannelId: string;
   ownerAuthorId: string;
@@ -191,6 +209,7 @@ interface UserConfig {
   lichess: UserLichessConfig;
   openai: UserOpenAiConfig;
   twitter: UserTwitterConfig;
+  db: UserDbConfig;
   discord: UserDiscordConfig;
   decapi: UserDecapiConfig;
   heartRate: UserHeartRateConfig;