+++ src/discord/discord.service.ts
@@ -83,14 +83,18 @@ export class DiscordService {
     }
   }
 
-  botSpeak(
+  async botSpeak(
     discordMessage: DiscordMessage | { channelId: string },
     message: string
   ) {
     const channel = this.client.channels.cache.get(
       discordMessage.channelId
     ) as TextChannel;
-    channel.send(message);
+    if (channel) {
+      await channel.send(message);
+    } else {
+      this.logger.error(`Channel not found: ${discordMessage.channelId}`);
+    }
   }
 
   // TODO: There's a bug sometimes where there's no message or context?
@@ -156,6 +160,8 @@ export class DiscordService {
     context.userId = discordMessage.author.id;
     context.isOwner =
       discordMessage.author.id === CONFIG.get().discord.ownerAuthorId;
+    context.isBot =
+      discordMessage.author.id === CONFIG.get().discord.botAuthorId;
     // We determine a mod by seeing if they're in the mod channel
     context.isMod =
       discordMessage.author.id === CONFIG.get().discord.ownerAuthorId ||