+++ src/models/base.entity.ts
@@ -1,0 +1,41 @@
+import {
+  Column,
+  UpdateDateColumn,
+  CreateDateColumn,
+  DeleteDateColumn,
+  VersionColumn,
+  PrimaryGeneratedColumn
+} from 'typeorm';
+
+export abstract class BaseEntity {
+  @PrimaryGeneratedColumn('increment')
+  id: number;
+
+  @VersionColumn()
+  version: number;
+
+  @CreateDateColumn({
+    type: 'timestamp',
+    name: 'created_at',
+    default: () => 'CURRENT_TIMESTAMP'
+  })
+  createdAt: Date;
+
+  @UpdateDateColumn({
+    type: 'timestamp',
+    name: 'updated_at',
+    default: () => 'CURRENT_TIMESTAMP'
+  })
+  updatedAt: Date;
+
+  @DeleteDateColumn({
+    type: 'timestamp',
+    name: 'deleted_at'
+  })
+  deletedAt: Date;
+
+  @Column({
+    type: 'json'
+  })
+  meta: JSON;
+}