+++ src/models/stored-command/stored-command.types.d.ts
@@ -1,0 +1,6 @@
+export interface StoredCommandInterface {
+  name: string;
+  message: string;
+}
+
+export interface StoredCommandCreateInput extends StoredCommandInterface {}